Move JetBrainsMinimumVersionChange to separate file
