+++ jetbrains-core/src/software/aws/toolkits/jetbrains/core/notification/NoticeType.kt
@@ -3,10 +3,7 @@
 
 package software.aws.toolkits.jetbrains.core.notification
 
-import com.intellij.openapi.application.ApplicationInfo
-import com.intellij.openapi.application.ApplicationNamesInfo
 import com.intellij.openapi.extensions.ExtensionPointName
-import software.aws.toolkits.resources.message
 
 interface NoticeType {
     val id: String
@@ -27,39 +24,6 @@ interface NoticeType {
 
         internal fun notices(): List<NoticeType> = EP_NAME.extensions.toList()
     }
-}
-
-class JetBrainsMinimumVersionChange : NoticeType {
-    override val id: String = "JetBrainsMinimumVersion_192"
-    private val noticeContents = NoticeContents(
-        message("notice.title.jetbrains.minimum.version.2019.2"),
-        message("notice.message.jetbrains.minimum.version.2019.2", ApplicationNamesInfo.getInstance().fullProductName)
-    )
-
-    override fun getSuppressNotificationValue(): String = ApplicationInfo.getInstance().fullVersion
-
-    override fun isNotificationSuppressed(previousSuppressNotificationValue: String?): Boolean {
-        previousSuppressNotificationValue?.let {
-            return previousSuppressNotificationValue == getSuppressNotificationValue()
-        }
-        return false
-    }
-
-    override fun isNotificationRequired(): Boolean {
-        val appInfo = ApplicationInfo.getInstance()
-        val majorVersion = appInfo.majorVersion.toIntOrNull()
-        val minorVersion = appInfo.minorVersion.toIntOrNull()
-
-        majorVersion?.let {
-            minorVersion?.let {
-                return majorVersion < 2019 || (majorVersion == 2019 && minorVersion < 2)
-            }
-        }
-
-        return true
-    }
-
-    override fun getNoticeContents(): NoticeContents = noticeContents
 }
 
 data class NoticeContents(val title: String, val message: String)
